+++ src/entities/task.entity.ts
@@ -27,7 +27,7 @@ export class Task {
   @Column({ type: 'boolean', default: true })
   status: boolean;
 
-  @Column({ name: 'estimated_time', type: 'bigint' })
+  @Column({ name: 'estimated_time', type: 'smallint' })
   estimatedTime: number;
 
   @Column({ name: 'is_public', type: 'boolean', default: false })
